requirements.txt
apipkg==1.5
asteval==0.9.1
backports.functools-lru-cache==1.6.1
coverage==3.7.1
cycler==0.10.0
DBUtils==1.1
enum34==1.1.10
execnet==1.7.1
funcsigs==1.0.2
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==2.2.5
mock==1.0.1
numpy==1.12.1
nupic==1.0.5
nupic.bindings==1.0.6
ordereddict==1.1
prettytable==0.7.2
psutil==1.0.1
py==1.8.1
pycapnp==0.6.3
PyMySQL==0.6.2
pyparsing==2.4.7
pyproj==1.9.3
pytest==3.0.7
pytest-cov==2.5.0
pytest-xdist==1.16.0
python-dateutil==2.1
pytz==2020.1
PyYAML==3.10
scikit-learn==0.20.4
scipy==1.2.3
singledispatch==3.4.0.3
six==1.15.0
subprocess32==3.5.4
tqdm==4.46.0
unittest2==0.5.1
validictory==0.9.1
